desktop/aiagent/windows_background_mode/aiagent/requirements.txt
requests
Pillow
pyinstaller
nuitka